Delivery data as of 2021-04-12
This includes the weekly delivery for vaccination centers from last
week (week number 14), as well as the first delivery to general
practitioners of ~940k BioNTech doses. Some of those 940k doses were
from this week's BioNTech delivery, and some of them were taken from the
special delivery for border regions that took place in the previous week
on 2021-03-30 (week number 13).

Apparently there were some AstraZeneca doses delivered to federal
authorities recently, but those are not included in this update since
distribution to states will happen in this week (week number 15).