Only show one chart at a time, add chart switcher
This is a bit easier to navigate and hopefully it's now clear which
charts support which filter options.

Also made it so there is now a line between data points with "gaps"
because only weekly data is available.

The code might be a bit easier to follow now, but still pretty ugly.